# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  title: Api
  version: 0.1.0
  description: Disaster Preparedness and Response Education System API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /users/me:
    get:
      operationId: getMe
      tags: [users]
      summary: Get current user profile
      responses:
        "200":
          description: Current user
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/UserProfile"
        "401":
          description: Unauthorized
    patch:
      operationId: updateMe
      tags: [users]
      summary: Update current user profile
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UserProfileUpdate"
      responses:
        "200":
          description: Updated user
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/UserProfile"
...
  /users/{id}/role:
    patch:
      operationId: updateUserRole
      tags: [users]
      summary: Update user role (admin only)
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UserRoleUpdate"
      responses:
        "200":
          description: Updated user
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/UserProfile"
...
  /quizzes:
    get:
      operationId: listQuizzes
      tags: [quizzes]
      summary: List all quizzes
      parameters:
        - name: category
          in: query
          required: false
          schema:
            type: string
      responses:
        "200":
          description: List of quizzes
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Quiz"
    post:
      operationId: createQuiz
      tags: [quizzes]
      summary: Create a quiz (teacher/admin)
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/QuizInput"
      responses:
        "201":
          description: Created quiz
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Quiz"
  /quizzes/{id}:
    get:
      operationId: getQuiz
      tags: [quizzes]
      summary: Get a quiz with questions
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Quiz with questions
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/QuizWithQuestions"
        "404":
          description: Not found
  /quizzes/{id}/submit:
    post:
      operationId: submitQuiz
      tags: [quizzes]
      summary: Submit quiz answers
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/QuizSubmission"
      responses:
        "200":
          description: Quiz result
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/QuizResult"
  /quizzes/{id}/questions:
    post:
      operationId: addQuestion
      tags: [quizzes]
      summary: Add a question to a quiz (teacher/admin)
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/QuestionInput"
      responses:
        "201":
          description: Created question
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Question"
...
components:
  schemas:
...
    UserProfile:
      type: object
      properties:
        id:
          type: integer
        clerkId:
          type: string
        name:
          type: string
        email:
          type: string
        role:
          type: string
          enum: [student, teacher, admin]
        createdAt:
          type: string
          format: date-time
      required:
        - id
        - clerkId
        - name
        - email
        - role
        - createdAt
    UserProfileUpdate:
      type: object
      properties:
        name:
          type: string
    UserRoleUpdate:
      type: object
      properties:
        role:
          type: string
          enum: [student, teacher, admin]
      required:
        - role
...
    Quiz:
      type: object
      properties:
        id:
          type: integer
        title:
          type: string
        description:
          type: string
        category:
          type: string
          enum: [earthquake, flood, fire, cyclone, pandemic, landslide, heatwave]
        questionCount:
          type: integer
        createdAt:
          type: string
          format: date-time
      required:
        - id
        - title
        - description
        - category
        - questionCount
        - createdAt
    QuizInput:
      type: object
      properties:
        title:
          type: string
        description:
          type: string
        category:
          type: string
          enum: [earthquake, flood, fire, cyclone, pandemic, landslide, heatwave]
      required:
        - title
        - description
        - category
    QuizWithQuestions:
      type: object
      properties:
        id:
          type: integer
        title:
          type: string
        description:
          type: string
        category:
          type: string
        questionCount:
          type: integer
        createdAt:
          type: string
          format: date-time
        questions:
          type: array
          items:
            $ref: "#/components/schemas/Question"
      required:
        - id
        - title
        - description
        - category
        - questionCount
        - createdAt
        - questions
    Question:
      type: object
      properties:
        id:
          type: integer
        quizId:
          type: integer
        questionText:
          type: string
        optionA:
          type: string
        optionB:
          type: string
        optionC:
          type: string
        optionD:
          type: string
        correctAnswer:
          type: ["string", "null"]
      required:
        - id
        - quizId
        - questionText
        - optionA
        - optionB
        - optionC
        - optionD
    QuestionInput:
      type: object
      properties:
        questionText:
          type: string
        optionA:
          type: string
        optionB:
          type: string
        optionC:
          type: string
        optionD:
          type: string
        correctAnswer:
          type: string
      required:
        - questionText
        - optionA
        - optionB
        - optionC
        - optionD
        - correctAnswer
    QuizSubmission:
      type: object
      properties:
        answers:
          type: array
          items:
            $ref: "#/components/schemas/QuizAnswer"
      required:
        - answers
    QuizAnswer:
      type: object
      properties:
        questionId:
          type: integer
        answer:
          type: string
      required:
        - questionId
        - answer
    QuizResult:
      type: object
      properties:
        id:
          type: integer
        userId:
          type: integer
        quizId:
          type: integer
        quizTitle:
          type: string
        score:
          type: integer
        totalQuestions:
          type: integer
        percentage:
          type: number
        completedAt:
          type: string
          format: date-time
      required:
        - id
        - userId
        - quizId
        - quizTitle
        - score
        - totalQuestions
        - percentage
        - completedAt
...